update to use common fabric base model
