+++ plugins/module_utils/models/manage_fabric/manage_fabric_ebgp.py
@@ -9,19 +9,16 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 import re
-from typing import List, Dict, Any, Optional, ClassVar, Literal
+from typing import List, Optional, ClassVar, Literal
 
-from ansible_collections.cisco.nd.plugins.module_utils.models.base import NDBaseModel
 from ansible_collections.cisco.nd.plugins.module_utils.models.nested import NDNestedModel
 from ansible_collections.cisco.nd.plugins.module_utils.common.pydantic_compat import (
     ConfigDict,
     Field,
     field_validator,
-    model_validator,
 )
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.enums import (
     AimlQosPolicyEnum,
-    AlertSuspendEnum,
     AllowVlanOnLeafTorPairingEnum,
     BgpAsModeEnum,
     BgpAuthenticationKeyTypeEnum,
@@ -32,7 +29,6 @@ from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.enum
     FabricTypeEnum,
     FirstHopRedundancyProtocolEnum,
     GreenfieldDebugFlagEnum,
-    LicenseTierEnum,
     MacsecAlgorithmEnum,
     MacsecCipherSuiteEnum,
     OverlayModeEnum,
@@ -40,8 +36,6 @@ from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.enum
     RendezvousPointCountEnum,
     RendezvousPointModeEnum,
     ReplicationModeEnum,
-    TelemetryCollectionTypeEnum,
-    TelemetryStreamingProtocolEnum,
     UnderlayMulticastGroupAddressLimitEnum,
     VpcPeerKeepAliveOptionEnum,
     VrfLiteAutoConfigEnum,
@@ -51,11 +45,9 @@ from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.enum
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.manage_fabric_common import (
     BGP_ASN_RE,
     BootstrapSubnetModel,
-    ExternalStreamingSettingsModel,
-    LocationModel,
     NetflowSettingsModel,
-    TelemetrySettingsModel,
 )
+from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.manage_fabric_base import FabricBaseModel
 
 """
 # Comprehensive Pydantic models for eBGP VXLAN fabric management via Nexus Dashboard
@@ -386,7 +378,7 @@ class VxlanEbgpManagementModel(NDNestedModel):
     pim_hello_authentication_key: str = Field(alias="pimHelloAuthenticationKey", description="3DES Encrypted", default="")
 
     # Management Settings
-    nxapi: bool = Field(description="Enable NX-API over HTTPS", default=False)
+    nxapi: bool = Field(description="Enable NX-API over HTTPS", default=True)
     nxapi_http: bool = Field(alias="nxapiHttp", description="Enable NX-API over HTTP", default=False)
     nxapi_https_port: int = Field(alias="nxapiHttpsPort", description="HTTPS port for NX-API", ge=1, le=65535, default=443)
     nxapi_http_port: int = Field(alias="nxapiHttpPort", description="HTTP port for NX-API", ge=1, le=65535, default=80)
@@ -802,7 +794,7 @@ class VxlanEbgpManagementModel(NDNestedModel):
         return value.lower()
 
 
-class FabricEbgpModel(NDBaseModel):
+class FabricEbgpModel(FabricBaseModel):
     """
     # Summary
 
@@ -814,82 +806,13 @@ class FabricEbgpModel(NDBaseModel):
     - `TypeError` - If field types don't match expected types
     """
 
-    model_config = ConfigDict(str_strip_whitespace=True, validate_assignment=True, populate_by_name=True, extra="allow")
-
-    identifiers: ClassVar[Optional[List[str]]] = ["fabric_name"]
-    identifier_strategy: ClassVar[Optional[Literal["single", "composite", "hierarchical", "singleton"]]] = "single"
-
-    # Basic Fabric Properties
-    category: Literal["fabric"] = Field(description="Resource category", default="fabric")
-    fabric_name: str = Field(alias="name", description="Fabric name", min_length=1, max_length=64)
-    location: Optional[LocationModel] = Field(description="Geographic location of the fabric", default=None)
-
-    # License, Telemetry, and Operations
-    license_tier: LicenseTierEnum = Field(alias="licenseTier", description="License Tier for fabric.", default=LicenseTierEnum.ESSENTIALS)
-    alert_suspend: AlertSuspendEnum = Field(
-        alias="alertSuspend", description="Alert Suspend state configured on the fabric.", default=AlertSuspendEnum.DISABLED
-    )
-    telemetry_collection: bool = Field(alias="telemetryCollection", description="Enable telemetry collection.", default=True)
-    telemetry_collection_type: TelemetryCollectionTypeEnum = Field(
-        alias="telemetryCollectionType", description="Telemetry collection method.", default=TelemetryCollectionTypeEnum.IN_BAND
-    )
-    telemetry_streaming_protocol: TelemetryStreamingProtocolEnum = Field(
-        alias="telemetryStreamingProtocol", description="Telemetry Streaming Protocol.", default=TelemetryStreamingProtocolEnum.IPV4
-    )
-    telemetry_source_interface: str = Field(
-        alias="telemetrySourceInterface",
-        description="Telemetry Source Interface Loopback ID, only valid if Telemetry Collection is set to inBand.",
-        default="loopback0",
-    )
-    telemetry_source_vrf: str = Field(
-        alias="telemetrySourceVrf", description="VRF over which telemetry is streamed, valid only if Telemetry Collection is set to inBand.", default="default"
-    )
-    security_domain: str = Field(alias="securityDomain", description="Security Domain associated with the fabric.", default="all")
+    _fabric_type: ClassVar[FabricTypeEnum] = FabricTypeEnum.VXLAN_EBGP
 
     # Core Management Configuration
     management: Optional[VxlanEbgpManagementModel] = Field(description="eBGP VXLAN management configuration", default=None)
 
-    # Optional Advanced Settings
-    telemetry_settings: Optional[TelemetrySettingsModel] = Field(alias="telemetrySettings", description="Telemetry configuration", default=None)
-    external_streaming_settings: ExternalStreamingSettingsModel = Field(
-        alias="externalStreamingSettings", description="External streaming settings", default_factory=ExternalStreamingSettingsModel
-    )
-
-    @field_validator("fabric_name")
-    @classmethod
-    def validate_fabric_name(cls, value: str) -> str:
-        """
-        # Summary
-
-        Validate fabric name format and characters.
-
-        ## Raises
-
-        - `ValueError` - If name contains invalid characters or format
-        """
-        if not re.match(r"^[a-zA-Z0-9_-]+$", value):
-            raise ValueError(f"Fabric name can only contain letters, numbers, underscores, and hyphens, got: {value}")
-        return value
-
-    @model_validator(mode="after")
-    def validate_fabric_consistency(self) -> "FabricEbgpModel":
-        """
-        # Summary
-
-        Validate consistency between fabric settings and management configuration.
-
-        ## Raises
-
-        - `ValueError` - If fabric settings are inconsistent
-        """
-        if self.management is not None and self.management.type != FabricTypeEnum.VXLAN_EBGP:
-            raise ValueError(f"Management type must be {FabricTypeEnum.VXLAN_EBGP}")
-
-        # Propagate fabric name to management model
-        if self.management is not None:
-            self.management.name = self.fabric_name
-
-        # Propagate BGP ASN to site_id if both are set and site_id is empty
+    def _post_validate_consistency(self) -> None:
+        """Propagate BGP ASN to site_id if site_id is empty."""
         if self.management is not None and self.management.site_id == "" and self.management.bgp_asn is not None:
             bgp_asn = self.management.bgp_asn
             if "." in bgp_asn:
@@ -897,32 +820,6 @@ class FabricEbgpModel(NDBaseModel):
                 self.management.site_id = str(int(high) * 65536 + int(low))
             else:
                 self.management.site_id = bgp_asn
-
-        # Auto-create default telemetry settings if collection is enabled
-        if self.telemetry_collection and self.telemetry_settings is None:
-            self.telemetry_settings = TelemetrySettingsModel()
-
-        return self
-
-    def to_diff_dict(self, **kwargs) -> Dict[str, Any]:
-        """Export for diff comparison, excluding fields that ND overrides for eBGP fabrics."""
-        d = super().to_diff_dict(**kwargs)
-        # ND always returns nxapiHttp=True for eBGP fabrics regardless of the configured value,
-        # so exclude it from diff comparison to prevent a persistent false-positive diff.
-        if "management" in d:
-            d["management"].pop("nxapiHttp", None)
-        return d
-
-    @classmethod
-    def get_argument_spec(cls) -> Dict:
-        return dict(
-            state={
-                "type": "str",
-                "default": "merged",
-                "choices": ["merged", "replaced", "deleted", "overridden"],
-            },
-            config={"required": False, "type": "list", "elements": "dict"},
-        )
 
 
 # Export all models for external use
