+++ plugins/module_utils/models/manage_fabric/manage_fabric_external.py
@@ -11,32 +11,24 @@ __metaclass__ = type
 import re
 from typing import List, Dict, Optional, ClassVar, Literal
 
-from ansible_collections.cisco.nd.plugins.module_utils.models.base import NDBaseModel
 from ansible_collections.cisco.nd.plugins.module_utils.models.nested import NDNestedModel
 from ansible_collections.cisco.nd.plugins.module_utils.common.pydantic_compat import (
     ConfigDict,
     Field,
     field_validator,
-    model_validator,
 )
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.enums import (
-    AlertSuspendEnum,
     CoppPolicyEnum,
     DhcpProtocolVersionEnum,
     FabricTypeEnum,
-    LicenseTierEnum,
     PowerRedundancyModeEnum,
-    TelemetryCollectionTypeEnum,
-    TelemetryStreamingProtocolEnum,
 )
 from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.manage_fabric_common import (
     BGP_ASN_RE,
     BootstrapSubnetModel,
-    ExternalStreamingSettingsModel,
-    LocationModel,
     NetflowSettingsModel,
-    TelemetrySettingsModel,
 )
+from ansible_collections.cisco.nd.plugins.module_utils.models.manage_fabric.manage_fabric_base import FabricBaseModel
 
 """
 # Comprehensive Pydantic models for External Connectivity fabric management via Nexus Dashboard
@@ -407,7 +399,7 @@ class ExternalConnectivityManagementModel(NDNestedModel):
         return value
 
 
-class FabricExternalConnectivityModel(NDBaseModel):
+class FabricExternalConnectivityModel(FabricBaseModel):
     """
     # Summary
 
@@ -422,103 +414,10 @@ class FabricExternalConnectivityModel(NDBaseModel):
     - `TypeError` - If field types don't match expected types
     """
 
-    model_config = ConfigDict(
-        str_strip_whitespace=True, validate_assignment=True, populate_by_name=True, extra="allow"  # Allow extra fields from API responses
-    )
-
-    identifiers: ClassVar[Optional[List[str]]] = ["fabric_name"]
-    identifier_strategy: ClassVar[Optional[Literal["single", "composite", "hierarchical", "singleton"]]] = "single"
-
-    # Basic Fabric Properties
-    category: Literal["fabric"] = Field(description="Resource category", default="fabric")
-    fabric_name: str = Field(alias="name", description="Fabric name", min_length=1, max_length=64)
-    location: Optional[LocationModel] = Field(description="Geographic location of the fabric", default=None)
-
-    # License, Telemetry, and Operations
-    license_tier: LicenseTierEnum = Field(alias="licenseTier", description="License Tier for fabric.", default=LicenseTierEnum.ESSENTIALS)
-    alert_suspend: AlertSuspendEnum = Field(
-        alias="alertSuspend", description="Alert Suspend state configured on the fabric.", default=AlertSuspendEnum.DISABLED
-    )
-    telemetry_collection: bool = Field(alias="telemetryCollection", description="Enable telemetry collection.", default=True)
-    telemetry_collection_type: TelemetryCollectionTypeEnum = Field(
-        alias="telemetryCollectionType", description="Telemetry collection method.", default=TelemetryCollectionTypeEnum.IN_BAND
-    )
-    telemetry_streaming_protocol: TelemetryStreamingProtocolEnum = Field(
-        alias="telemetryStreamingProtocol", description="Telemetry Streaming Protocol.", default=TelemetryStreamingProtocolEnum.IPV4
-    )
-    telemetry_source_interface: str = Field(
-        alias="telemetrySourceInterface",
-        description="Telemetry Source Interface Loopback ID, only valid if Telemetry Collection is set to inBand.",
-        default="loopback0",
-    )
-    telemetry_source_vrf: str = Field(
-        alias="telemetrySourceVrf", description="VRF over which telemetry is streamed, valid only if Telemetry Collection is set to inBand.", default="default"
-    )
-    security_domain: str = Field(alias="securityDomain", description="Security Domain associated with the fabric.", default="all")
+    _fabric_type: ClassVar[FabricTypeEnum] = FabricTypeEnum.EXTERNAL_CONNECTIVITY
 
     # Core Management Configuration
     management: Optional[ExternalConnectivityManagementModel] = Field(description="External Connectivity management configuration", default=None)
-
-    # Optional Advanced Settings
-    telemetry_settings: Optional[TelemetrySettingsModel] = Field(alias="telemetrySettings", description="Telemetry configuration", default=None)
-    external_streaming_settings: ExternalStreamingSettingsModel = Field(
-        alias="externalStreamingSettings", description="External streaming settings", default_factory=ExternalStreamingSettingsModel
-    )
-
-    @field_validator("fabric_name")
-    @classmethod
-    def validate_fabric_name(cls, value: str) -> str:
-        """
-        # Summary
-
-        Validate fabric name format and characters.
-
-        ## Raises
-
-        - `ValueError` - If name contains invalid characters or format
-        """
-        if not re.match(r"^[a-zA-Z0-9_-]+$", value):
-            raise ValueError(f"Fabric name can only contain letters, numbers, underscores, and hyphens, got: {value}")
-
-        return value
-
-    @model_validator(mode="after")
-    def validate_fabric_consistency(self) -> "FabricExternalConnectivityModel":
-        """
-        # Summary
-
-        Validate consistency between fabric settings and management configuration.
-
-        ## Raises
-
-        - `ValueError` - If fabric settings are inconsistent
-        """
-        # Ensure management type matches model type
-        if self.management is not None and self.management.type != FabricTypeEnum.EXTERNAL_CONNECTIVITY:
-            raise ValueError(f"Management type must be {FabricTypeEnum.EXTERNAL_CONNECTIVITY}")
-
-        # Propagate fabric name to management model
-        if self.management is not None:
-            self.management.name = self.fabric_name
-
-        # Validate telemetry consistency
-        if self.telemetry_collection and self.telemetry_settings is None:
-            # Auto-create default telemetry settings if collection is enabled
-            self.telemetry_settings = TelemetrySettingsModel()
-
-        return self
-
-    # TODO: to generate from Fields (low priority)
-    @classmethod
-    def get_argument_spec(cls) -> Dict:
-        return dict(
-            state={
-                "type": "str",
-                "default": "merged",
-                "choices": ["merged", "replaced", "deleted", "overridden"],
-            },
-            config={"required": False, "type": "list", "elements": "dict"},
-        )
 
 
 # Export all models for external use
